CB-77 Implement Order Table

+++ src/main/java/com/ooadprojectserver/restaurantmanagement/repository/order/OrderRepository.java
@@ -10,6 +10,6 @@ import java.util.UUID;
 public interface OrderRepository extends JpaRepository<Order, UUID> {
 
     @Query("select o from Order o where o.staff.id = ?1 and o.orderTable.id = ?2 and o.status = ?3")
-    Order findOrder(UUID id, Long table_id, OrderStatus status);
+    Order findOrder(UUID id, Integer table_id, OrderStatus status);
 
 }
